Change de structure

+++ blablacar-feeder/src/main/java/es/ulpgc/dacd/domain/Trip.java
@@ -47,7 +47,7 @@ public final class Trip {
         return arrival;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
